+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+No unreleased changes at this point.
+
+## [0.2.1] - 2020-12-16
+
 ## Added
 - Use parallel builds for ClojureScript compilation
 
@@ -72,7 +76,8 @@ extern file in location `src/<project-dirs>/client/externs.js`.
 - **Breaking change** - Now by default this module doesn't provide example api integrant key.
 To retain it, you have to add `:add-example-api? true` to the config options.
 
-[UNRELEASED]:  https://github.com/magnetcoop/hydrogen.module.core/compare/v0.2.0...HEAD
+[UNRELEASED]:  https://github.com/magnetcoop/hydrogen.module.core/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/magnetcoop/hydrogen.module.core/releases/tag/v0.2.1
 [0.2.0]: https://github.com/magnetcoop/hydrogen.module.core/releases/tag/v0.2.0
 [0.1.11]: https://github.com/magnetcoop/hydrogen.module.core/releases/tag/v0.1.11
 [0.1.9]: https://github.com/magnetcoop/hydrogen.module.core/releases/tag/v0.1.9